poker_bot/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiosqlite==0.20.0
treys==0.1.8
openai>=1.12.0
python-dotenv==1.0.1
aiogram==3.3.0
aiofiles==23.2.1
